nekodigi/streamlit2/requirements.txt
streamlit>=1.41
pillow>=11
numpy>=2
opencv-python-headless>=4.10
openai>=1.60
firebase-admin>=6
